Reduce utils/kebab-case.ts to utils/inflector.ts
