+++ packages/generator/src/generators/page-generator.ts
@@ -1,6 +1,6 @@
 import {join} from "path"
 import {Generator, GeneratorOptions} from "../generator"
-import {camelCaseToKebabCase} from "../utils/kebab-case"
+import {camelCaseToKebabCase} from "../utils/inflector"
 
 export interface PageGeneratorOptions extends GeneratorOptions {
   ModelName: string
