+++ packages/generator/src/utils/inflector.ts
@@ -21,3 +21,7 @@ export const singlePascal = pipe(singular, capitalize)
 export const singleCamel = pipe(singular, uncapitalize)
 export const pluralPascal = pipe(plural, capitalize)
 export const pluralCamel = pipe(plural, uncapitalize)
+
+export function camelCaseToKebabCase(transformString: string) {
+  return transformString.replace(/([a-z0-9]|(?=[A-Z]))([A-Z])/g, "$1-$2").toLowerCase()
+}
