--- packages/generator/src/utils/kebab-case.ts
@@ -1,3 +1,0 @@
-export function camelCaseToKebabCase(transformString: string) {
-  return transformString.replace(/([a-z0-9]|(?=[A-Z]))([A-Z])/g, "$1-$2").toLowerCase()
-}
